extract handlers{,_test}.go
